requirements.txt
# Основные зависимости
flask==2.3.3
requests==2.31.0

# Aiogram
aiogram==3.10.0
aiohttp==3.9.1

# Environment variables
python-dotenv==1.0.0
